Initial ip_fragmentation handling
IP Fragmentation in eBPF using deticated hashmap storing dst_addr by
{saddr;daddr;id;protocol}. Current implementation doesnt expires entries
since we need to extend our current bpf code to support bpf_timer. bpf_timer
need BTF so instead of re-inventing the wheel we are going port current code
to support libbpf.

The current implementation doesnt take ordering into account, but in most mobile
core-network this can be an issue since most of the time ip_fragmentation
is done on service cards that differ from data-plane or are not using the same
CPU nor the same emitting interface for all the fragments. If first frag received
is not the first ip_frag then we simply dont have GTP header to fetch TEID to know
remote endpoint for rewriting ip_header.daddr. So there is a need to store until
very first segment is received. This is typically a limitation for eBPF program
since working in driver context you just have a limited space for storing
payload and then send burst of packet. One solution would be to use AF_XDP for
this particular use-case.

So we are preparing the road for futur work on XDP by incorporating libbpf from
official kernel mirror.