Resolver: Rewrote pGW scheduling selection
Scheduling algo is built around 3 stages :

(1) APN configuration reference's a list of "service-selection". Each one can
    define a priority. While scheduling NAPTR service from DNS, entries are
    matched according to this priority. Smallest priority defines higher
    scheduling priority.

(2) The same service can be used for multiple DNS NAPTR records. If so
    3GPP NAPTR define's DNS "Order" to specify selection priority. The
    smallest NAPTR Order entry defines the higher priority.

(3) DNS NAPTR can returns A or SRV record. A SRV record can reference a list of
    A record, each one can define a weight. In this last case we are using
    weight attribute in order to drive a WLC scheduling decision.

Note: In this last stage, a "priority" attribute is available but we are
simply not taking it into account, since weight scheduling and this 3 stages
algo sound already enough complexity.