test: build a topo and perform a GTPc ping
example of usage:
sudo ./testenv.sh \
  -i /home/vjardin/dev/gtping/src/gtping \
  -g /home/vjardin/dev/gtp-guard/bin/gtp-guard \
  -c /tmp/gtp-guard.conf \
  -f /home/vjardin/dev/gtp-guard/src/bpf/gtp_fwd.bpf \
  -r /home/vjardin/dev/gtp-guard/src/bpf/gtp_route.bpf \
  -m /home/vjardin/dev/gtp-guard/src/bpf/gtp_mirror.bpf \
  -k no

It creates 3 netns and 1 sandbox one.

The sandbox netns is used to check eBPF modules.

The 3 netns are used as a virtual topology in order to have simulated
end to end setup.

This script is designed to be run locally or from github's Actions.