make: CFLAGS with Werror - next
Since all the warnings are sanitized, let's stop the compilation on any
warning. The main benefit is when the github actions are running,
they'll stop during the checks of a pull request.

I forgot to include it with the commit e0d2d6e29785c06597bfe9151f473fe21e05ba71