pppoe: Add support to virtual ethernet address per gtp session
A virtual ethernet address is generated based on IMSI, but this
address MUST be unique and different for all other gtp session
for this same IMSI.

As per RFC5342.2.1 Local bit is set. First byte of ethernet address
is set according to the rule : hbits lbits=0000 0010
It's offering 3bits for unique ethernet address : 8 possibilities
hbits are reserved for upcoming feature.