switch: Add support to "gtpc-egress-tunnel-endpoint"
This feature allow a split in GTP-C when used in switching mode (proxying).
It will create a socket pair with ingress & egress GTP-C traffic. Egress
traffic in switch mode is related to traffic leaving GTP-Guard to remote
pGW. This socket pair is then used for GTP-C relaying: when ingress GTP-C
is received then egress socket in socket pair is used for relaying to
remote pGW.

This feature allow ingress/egress traffic differentiation so that you can
fined option for GTP-C mirroring.