ci: add arm64
WIP

Meanwhile, even if 1 single OS is used, it could be extended later to
run the CI for some other Linux distributions. Frankly, I do not foresee the benefits,
but currently, let's keep it has a placeholder.

TODO: change the paths of the artifact for the tests to be run.