resolv: Add support to local bind while connecting remote nameserver.
This feature is available via following VTY command in apn conf :
"nameserver-bind (A.B.C.D|X:X:X:X) port <1024-65535>"

This feature is useful when your local networking configuration
is based on loopback interfaces (dummy). In that situation
connected interface and service interface are different so you
need to bind to the proper interface for service.

glibc resolver interface (not documented), is providing context
structure in order to set remote nameserver and connection to
use. We are then, creating our nameserver resolv socket directly,
bind to the proper Address & Port, and pass it to glibc as
pre-init.