add few gtpc, ppp and CLI logs
log gtpc_delete_session_request_hdl

When a eNodeB detects a radio error (for instance a cell phone switching
to plane mode), the MME or AMF emits a delete session request to the SGW
and then the SGW emits a GTPc delete session request to the PGW
(gtp-guard). Let's log this event.

ppp: add IPCP DNS logs
and fix display order for the IPCP IP Address Format.

CLI: show gtp session 0 (all)
so it dumps all the gtp session when session is 0