cdr: Support file rolling
When a CDR is added into the file and if the file size exceed max size
or roll time is reached then close current file and create a new one.

During close operation, truncation is performed to match header len.

In case of a crash, during restart, last current file will be used
and cdr will be appended into it.

Rolled file are automatically moved to archive directory, if configured.
If no archive directory is configured then fallback to document-root is
used.

CDR file sequence number is supported as a mitigation to any flooding.
In such a case we only support 8bits seq_num which lead to 256 files
and is used as a circular buffer. This protection prevent loose of
CDR when file are created at the same seconds.

Signed-off-by: Alexandre Cassen <acassen@gmail.com>