GTPv1: exclusive Sqn tracking for ingress messages
GTP-C Sollicited messages are refered to as GTP_INGRESS. Makes Sqn
update exclusive to these messages since it may exist some GTPv1
protocol anomaly or bogus implementation (you choose), where GTP-C TEID
IE is present in response message where it should not or mainly there is
no reason for. Update-PDP-Response for example can add GTP-C TEID
but there is no reasons for it since teid present in GTP header can be
used for that purpose. Mainly it is creating redundant information
leading to protocol inconsitence. So make the code more tolerant in
such a case.

Signed-off-by: Alexandre Cassen <acassen@gmail.com>