lib/thread: add generic thread_del
Thread lifetime could be summarized in these three steps:
 1. waiting: thread is added, waiting for event to happen (timer or fd)
 2. ready: fetched in ready list, not yet processed
 3. processing: we are in thread->func callback

After processing, thread is free/recycled from lauch_thread_scheduler()
loop, so there is normally no need to delete it from app.

Sometime it is desirable to manually delete them before event fire, or
when event is fired. The common case is to close associated fd, as the
thread-lib has to epoll_ctl(delete) before fd is closed.

Before, they were 4 functions for this job:
 - thread_cancel(): handles cases 1 and 2
 - thread_del_write(): handle case 3 for write fd.
 - thread_del_read(): handle case 3 for read fd.
 - thread_del_timer(): handle case 1 for timer thread.

If a function is called for an unsupported case, it results in an
application crash. Also, thread_cancel() does not support case 2 for
timer threads.

This patch gather these functions into thread_del(), which handle all
cases correclty.

Signed-off-by: Olivier Gournet <gournet.olivier@gmail.com>