replace build system from old plain Makefile to meson/ninja
Add a meson build system, that have a better handling of dependencies.

Change root Makefile so the old build system behavior is kept, so
commands will works the same:
  - make
  - make clean
  - make install
  - make tarball

Add command to build and run executable in tools/selftests
  - make test

System package libbpf is used if found and version >= 1.5.0, otherwise
libbpf is fetched and built as subproject from git (v1.6.1).

Signed-off-by: Olivier Gournet <gournet.olivier@gmail.com>