table: Support table_vty_out()
The table_format() function relies on a fixed-size buffer, and
truncation occurs if the formatted output exceeds the buffer
size.

Introduce table_vty_out() to iterate over the rows and display
each one to the VTY. There is no need to know how much memory
is needed to dump the entire table because iteration makes it
dynamic.

Signed-off-by: Alexandre Cassen <acassen@gmail.com>